feature(all): need to push (project should be done but 403 still)

+++ src/main/resources/sql/script.sql
@@ -5,7 +5,7 @@ USE `chatop`;
 -- Création de la table USERS
 DROP TABLE IF EXISTS `USERS`;
 CREATE TABLE IF NOT EXISTS `USERS` (
-                                       `id` INT PRIMARY KEY AUTO_INCREMENT,
+                                       `id` BIGINT PRIMARY KEY AUTO_INCREMENT,
                                        `email` varchar(255) NOT NULL,
                                        `name` varchar(255),
                                        `password` varchar(255),
@@ -16,13 +16,13 @@ CREATE TABLE IF NOT EXISTS `USERS` (
 -- Création de la table RENTALS
 DROP TABLE IF EXISTS `RENTALS`;
 CREATE TABLE IF NOT EXISTS `RENTALS` (
-                                         `id` INT PRIMARY KEY AUTO_INCREMENT,
+                                         `id` BIGINT PRIMARY KEY AUTO_INCREMENT,
                                          `name` varchar(255),
                                          `surface` DECIMAL(10, 2),
                                          `price` DECIMAL(10, 2),
                                          `picture` varchar(255),
                                          `description` varchar(2000),
-                                         `owner_id` INT NOT NULL,
+                                         `owner_id` BIGINT NOT NULL,
                                          `created_at` timestamp DEFAULT CURRENT_TIMESTAMP,
                                          `updated_at` timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
                                          CONSTRAINT `fk_owner` FOREIGN KEY (`owner_id`) REFERENCES `USERS` (`id`) ON DELETE CASCADE
@@ -31,9 +31,9 @@ CREATE TABLE IF NOT EXISTS `RENTALS` (
 -- Création de la table MESSAGES
 DROP TABLE IF EXISTS `MESSAGES`;
 CREATE TABLE IF NOT EXISTS `MESSAGES` (
-                                          `id` INT PRIMARY KEY AUTO_INCREMENT,
-                                          `rental_id` INT,
-                                          `user_id` INT,
+                                          `id` BIGINT PRIMARY KEY AUTO_INCREMENT,
+                                          `rental_id` BIGINT,
+                                          `user_id` BIGINT,
                                           `message` varchar(2000),
                                           `created_at` timestamp DEFAULT CURRENT_TIMESTAMP,
                                           `updated_at` timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
